Partial revert of [3133].  The NPEs were caused by another race condition.  The path of the arguments passed to the IComponent/ComponentModel comparators should never be null.
However, the TableModel.getFocusableComponentModel() method did not check for the possibility that the origin of the specified search vector is not currently part of a TableModel (i.e. it's path is null).  The fact that the origin has a null path is caused by the fact that various locations within TableView did not synchronize around multiple TableModel method calls to ensure an atomic operation.  In those cases, the focused ComponentModel was retrieved and then passed to TableModel.getFocusableComponentModel().  In between those calls, the focused ComponentModel was removed from the TableModel, causing its path to be null when called within getFocusableComponentModel().

So we now 1) assert that the paths of the arguments passed to the Comparators are not null, 2) ignore the search vector origin if it has a null path (i.e. the same as if the user passed a null search vector), and 3) synchronize blocks within TableView that should be atomic with respect to the TableEnvironmentModel.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3134 897c5806-f142-0410-a293-e1bf5698dec8