Changed the contract of the TableModel.getComponentModel(ComponentPath) method such that it can return null if a component model does not exist at the specified component path instead of throwing IllegalArgumentException.  This is similar to the change that was done for ITable.getComponent(ComponentPath) in [3119].
The purpose of this change is to accommodate a LEGAL race condition that the UI model layer must be able to handle.  An explanation of this race condition follows.

While we now guarantee that table events are serialized and delivered in the correct order, not everything results in an event being fired in the first place.  For example, during a drag operation, the drag components are added to a temporary container that is not part of the table when the components are added so no events are fired.  Then, when the container is added to the table, a single add event is fired and the UI model attempts to add the new container and all of its descendants.  HOWEVER, by the time the container added event is received, its components have been removed, so there is nothing to add.  At a later time, we get the component removed events, but those components were never added to the UI model so they don't exist.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3138 897c5806-f142-0410-a293-e1bf5698dec8