Fix for regression introduced in [3126].
The network node layer previously assumed that events fired by its local table instance are delivered synchronously.  Specifically, the LocalNetworkTable class uses the boolean ignoreEvents_ field to suppress processing of events received when it is the originator of the change that caused the event in the first place.

Because events are now delivered asynchronously, numerous warnings are written to the log during a network session that the player trying to modify the table is not the editor.  This is purely due to a race between the ignoreEvents_ field being set and the delivery of the events associated with the changes.

The employed fix is to replace the ignoreEvents_ field with a better test to determine if the event was originated by the network node layer as opposed to the user manipulating the local table via the UI.  We now associate the thread on which the event originated with instances of ComponentEvent and ContainerEvent.  The LocalNetworkTable class can then ignore any events it receives from the local table if the thread stored in the event instance equals the network node layer thread.

We've used this type of approach before where we are storing a snapshot of some state at the time the table was modified in the events fired by the table (e.g. the path to the component is a snapshot just in case the component has been moved by the time the listener receives the event).  So, while this specific change has a bit of a smell to it, it's not completely outside of what the perfect solution should be.  I think the long term solution will be to always associate certain table/component state with all events.  In that case, instead of using the thread that caused the change as the context, we may store some well-known ThreadLocals in the event at the time the change was made.  For example, the name of the player that made the change.  That would provide the required information for the LocalNetworkTable to make the decision to ignore or process a particular event.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3148 897c5806-f142-0410-a293-e1bf5698dec8