Fixed bug in ContainerModel that prevented deletion of child ComponentModels (which could also be ContainerModels).  If an IContainer is deleted from an ITable, a component removed event is only fired for the container, not for its descendants.  The ContainerModel.deleteComponentModel() method always (incorrectly) assumed that it didn't have to recursively delete its descendants because it assumed a corresponding component removed event would have been fired prior to it being deleted.
A lingering ComponentModel/ContainerModel never unregisters its IComponent/IContainer listeners, and so will continue to receive events even if the model is no longer part of a TableModel.  This leads to assertion failures in other parts of the code (namely, ContainerModel.ContainerModelListener.componentRemoved) where it is assumed the component model is actively part of a table model.

We are only noticing this legacy bug now because we previously could not drag and drop a container onto another container (other than the tabletop).

The fix is to recursively delete all child component models in ContainerModel.deleteComponentModel().

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3152 897c5806-f142-0410-a293-e1bf5698dec8