Fix for regression introduced in [3126].
Because table events are now delivered asynchronously, the TableModel.isDirty flag is not properly updated.  For example, when opening a new table file, the table is modified and then the isDirty flag is set to false on the UI thread.  However, all the events from modifying the table are fired on a separate thread, and most likely arrive after isDirty is set to false.  When each of these events arrive, it appears as if someone has modified the table after loading it, thus isDirty is set to true.

The fix is to associate a monotonically-increasing revision number with the table, and to have the TableModel record the revision number at which the table changes were last committed.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3154 897c5806-f142-0410-a293-e1bf5698dec8