#81
Allow user to override default collection of components to be dragged.  The default behavior is to delegate to the target component's drag strategy.  However, the user may override that behavior using the keyboard modifiers below:

CTRL        : drag only the target component
ALT         : drag the target component and all its siblings with a higher z-order
SHIFT + ALT : drag the target component and all its siblings with a lower z-order


git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3156 897c5806-f142-0410-a293-e1bf5698dec8