Fix test break in Maven build caused by r3269.
The test cases extended by those fixtures require an implementation of IComponentSurfaceDesignRegistry to be installed in the service registry.  Since the org.gamegineer.table.core bundle no longer provides an implementation of this interface, we need to ensure one is installed.  I simply modified the test runner to activate the org.gamegineer.table.core.impl bundle where the registry implementation now resides.

This kind of smells bad (why should the API bundle tests depend on a particular implementation of that API running in the framework?).  If we get pushback, we'll have to provide a fake implementation (and register it with DS) as part of the org.gamegineer.table.core.tests fragment.

git-svn-id: https://svn.code.sf.net/p/gamegineer/code/dev/trunk@3270 897c5806-f142-0410-a293-e1bf5698dec8