Fix SCM information.
Note that the Git SCM URIs in the generated site documentation will not be correct.  This is due to a limitation of the Maven Site plugin in which it assumes that all SCMs support hierarchical URIs when building a multi-module site.  Git does not directly support referencing subtrees as SVN does.  Thus, the generated SCMs will have extra path information following the repository URI, which must be removed to obtain a well-formed Git URI.

However, the web URIs are correct because GitHub does support hierarchical references via their web interface.