Suppress null analysis warning.
False positive due to Eclipse bug 477719.  Note that we need to suppress
"all" warnings because of some strange behavior with the analysis
engine.  It keeps oscillating between warning about unnecessary casts
and a required cast.  We'll reevaluate after 477719 is fixed to see if
that helps lock down the engine's analysis.