Provide external null annotations to compiler during Maven build
With the upgrade to Tycho 1.0.0, project-specific compiler settings are
now enabled by default.  Thus, null analysis is being performed.
However, the compiler does not have access to the external null
annotations we've defined, so numerous null analysis violations are
reported.

This change is only a temporary fix for two reasons.

First, it will only work when running the build in a specific manner,
which matches the manner in which the build runs on the CI server.  From
the repo root, building the entire main project, as in:

$ mvn -f main/org.gamegineer.main/pom.xml clean install

should result in no null analysis violations.  However, running the
build locally from any subproject will produce null analysis warnings as
the relative path to the external annotations is no longer correct.

Second, referring to files outside of the project is a Maven
anti-pattern.  The correct way to access the external annotations from
any project is to publish them to a Maven repository and configure them
as a dependency for all relevant projects.  That dependency can then be
unpacked locally and referred to in a project-agnostic manner.

An issue will be created to improve this implementation.  Our concern
right now is simply to get rid of the false null analysis warnings in
the CI build.